backend/requirements.txt
# =============================================================================
# Backend requirements
# =============================================================================

# --- Web framework ---
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1

# --- Validation ---
pydantic==2.7.1

# --- Core dependencies ---
opencv-python-headless==4.9.0.80
Pillow==10.3.0
nltk==3.8.1
spacy==3.7.4
rapidfuzz==3.9.3
sentence-transformers==2.7.0
scikit-learn==1.4.2
numpy==1.26.4

# --- PDF processing ---
pdfplumber==0.11.0
pdf2image==1.17.0

# --- LLM (Groq OpenAI-compatible) ---
openai>=1.40.0

# --- Testing ---
pytest==8.2.0
